+++ client/src/pages/SignUp.js
@@ -7,7 +7,7 @@ import { Typography , Grid} from "@material-ui/core";
 import TextField from '@material-ui/core/TextField';
 import SubmitButton from '../features/SubmitButton';
 import UserInputContainer from '../features/UserInputContainer';
-import NavBar from '../features/NavBar'
+import NavBar from '../features/NavBar/MainBody'
 
 const useStyles = makeStyles(theme => ({
   signUpText: {
@@ -30,8 +30,8 @@ const useStyles = makeStyles(theme => ({
   button: {
     margin: 10,
   },
-  root: {
-    height: 520,
+  container: {
+    height: 530,
     marginTop: 150,
   }
 }));
@@ -74,17 +74,20 @@ const Register = () => {
       if (!data.access_token) {
         alert(data.message);
       } else {
-        localStorage.setItem("token", data.access_token)
-        alert(data.message)
+        localStorage.setItem("mailsender_token", data.access_token);
+        alert(data.message);
       }
     })
+    .catch(err => {
+      console.log(err.message);
+    });
   }
 
   return (
     <Fragment>
       <NavBar />
       <UserInputContainer classes={{
-        root: classes.root
+        root: classes.container
       }} maxWidth="sm">
         <Typography className={classes.signUpText}>Sign up</Typography>
         <ValidatorForm
